==> 2.A.ADD_FEATURES_BY_KNOWN_IDS.ipynb <==
import pandas as pd
import numpy as np
import json
import datetime
import time

import config
import request_by_IDS

# import the labeled data
df = pd.read_csv('all_labeled_data_without_maryland.csv')
df.shape

df.head()

tweet_list = list(df['tweet_id'])

def recover_tweets(tweet_list, start_index, request_byIDS=request_by_IDS):
    """
    given a list of tweets call for the tweet from the endpoint in request_by_ID
    advance tweet by tweet until rate limited -- delay?

    return:
        - new_index:
        - collected_tweets: list of tweet objects
    """

    call_tweets = True
    new_index = start_index
    collected_tweets = []

    while call_tweets == True:

        try:
            twt = tweet_list[new_index]
            new_tweet = request_by_IDS.main(twt)
            collected_tweets.append(new_tweet)
            new_index += 1
            print(new_index)
        except:
            print(f"Last index used: {new_index}")
            call_tweets = False

    return new_index, collected_tweets

def clean_tweets_for_errors(tweet_list):
    output_tweets = []
    for tweet in tweet_list:
        if 'data' in tweet.keys():
            output_tweets.append(tweet['data'])
    return output_tweets

def create_tweets_df(raw_list):
    """
    given the list of tweet objects return a single data frame
    """
    list_of_tweets = []
    for tweet in raw_list:
        current_df = pd.DataFrame([tweet])
        list_of_tweets.append(current_df)
    df = pd.concat(list_of_tweets, ignore_index=True)
    return df

i, twts = recover_tweets(tweet_list, 0)
good_tweets = clean_tweets_for_errors(twts)
df_1  = create_tweets_df(good_tweets)

df_1.shape

%%time
i, twts = recover_tweets(tweet_list, 300)
good_tweets = clean_tweets_for_errors(twts)
df_2  = create_tweets_df(good_tweets)
time_end = datetime.datetime.now()

%%time
i, twts = recover_tweets(tweet_list, 600)
good_tweets = clean_tweets_for_errors(twts)
df_3  = create_tweets_df(good_tweets)
time_end = datetime.datetime.now()

%%time
i, twts = recover_tweets(tweet_list, 900)
good_tweets = clean_tweets_for_errors(twts)
df_4  = create_tweets_df(good_tweets)
time_end = datetime.datetime.now()

%%time
i, twts = recover_tweets(tweet_list, 1200)
good_tweets = clean_tweets_for_errors(twts)
df_5  = create_tweets_df(good_tweets)
time_end = datetime.datetime.now()
