==> 2.PRODUCE_FULL_LABELED_DATA.ipynb <==
import pandas as pd
import os

# parent event set folder from: 

all_files = []

hurricane_directories = [('hurricane_maria_2017', 'Maria'),
 ('hurricane_irma_2017', 'Irma'),
 ('hurricane_harvey_2017', 'Harvey'),
 ('hurricane_matthew_2016', 'Matthew')]

for d, name in hurricane_directories:
    files = os.listdir(d)
    for f in files:
        print(f)
        df = pd.read_csv(f"{d}/{f}", delimiter='\t')
        df['hurricane'] = name
        all_files.append(df)

len(all_files)

df = pd.concat(all_files)
df = df.drop_duplicates()

df.shape

df.head()

# df.to_csv('all_labeled_data_without_maryland.csv', index=False)

df_classes = df.groupby('class_label')['tweet_text'].count()
df_classes

df_2 = pd.read_csv('all_labeled_data.csv')
df_2.shape

df_2_classes = df_2.groupby('class_label')['tweet_text'].count()
df_2_classes

df.shape

df.head()

df.to_csv('website_all_labeled_2016_17_hurricanes.csv', index=False)
