==> 3.PREDICT_YES_NO_HURRICANE.ipynb <==
import pandas as pd
import numpy as np
import pickle

import re
import string
import contractions
import collections

import nltk
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize 
from nltk.stem import WordNetLemmatizer

from sklearn.feature_extraction.text import TfidfVectorizer

# load the nltk specific parameters
stop_words = nltk.corpus.stopwords.words('english')

# import trained text vectorizer
with open('./WINNING_MODEL/training_vectorizer.pkl', 'rb') as vec_file:
    vectorizer = pickle.load(vec_file)

# import trained classifier:
with open('./WINNING_MODEL/yes_no_svc_plus.pkl', 'rb') as file:
    svc = pickle.load(file)

# csv files from the API

# df = pd.read_csv('./ADD_HUR_DATA/sandy_2012_3day_increment_v2.csv')
# df.shape

# df = pd.read_csv('./ADD_HUR_DATA/arthur_2014_3day_increment_v2.csv')
# df.shape

# df = pd.read_csv('./ADD_HUR_DATA/barry_2019_3day_increment.csv')
# df.shape

# df = pd.read_csv('./ADD_HUR_DATA/florence_2018_3day_increment.csv')
# df.shape

# df = pd.read_csv('./ADD_HUR_DATA/isaias_2020_3day_increment.csv')
# df.shape

# df = pd.read_csv('./ADD_HUR_DATA/nicholas_2021_3day_increment.csv')
# df.shape

df = pd.read_csv('./ADD_HUR_DATA/hurricane_ian_3day_increment2.csv')
df.shape

# select or combine lemmatization, stemming, stopword removal to compare performance

def preprocess_text(text, flg_stemm = True, flg_lemm = True, lst_stopwords=None):
    text_clean = re.sub(r'[^\w\s.,]', '', str(text).strip())
    lst_text = text_clean.split()
    if lst_stopwords is not None:
        lst_text = [word for word in lst_text if word not in lst_stopwords]
    if flg_stemm == True:
        stemm = nltk.stem.porter.PorterStemmer()
        lst_text = [stemm.stem(word) for word in lst_text]
    if flg_lemm == True:
        lem=nltk.stem.wordnet.WordNetLemmatizer()
        lst_text = [lem.lemmatize(word) for word in lst_text]

    text_clean = ' '.join(filter(None, lst_text))
    text_clean = text_clean.replace(" ,",",").replace(' .', '.')
    text_clean = contractions.fix(text_clean)
    return text_clean

X = df['text'].apply(lambda x: preprocess_text(x, flg_stemm=True, flg_lemm=True, lst_stopwords=stop_words))
X.shape

# transform the target tweets
X_vec = vectorizer.transform(X)

preds = svc.predict(X_vec)

preds.shape

df['predictions'] = preds
df.shape

final_df = df[df['predictions'] == 1]
final_df.shape

final_df.tail()

# final_sandy.to_csv('./CLEAN_ADD_HUR_DATA/sandy_2012_yn.csv', index=False)

# final_df.to_csv('./CLEAN_ADD_HUR_DATA/arthur_2014_yn.csv', index=False)

# final_df.to_csv('./CLEAN_ADD_HUR_DATA/barry_2019_yn.csv', index=False)

# final_df.to_csv('./CLEAN_ADD_HUR_DATA/florence_2018_yn.csv', index=False)

# final_df.to_csv('./CLEAN_ADD_HUR_DATA/isaias_2020_yn.csv', index=False)

final_df.to_csv('./CLEAN_ADD_HUR_DATA/ian_2022_yn.csv', index=False)
