==> 4.TRAIN_DATA_ADD_NAME_NORMALIZE_DATE.ipynb <==
import pandas as pd
import numpy as np
import json
import tweepy
import config
from sklearn.preprocessing import MinMaxScaler

df_l = pd.read_csv('labeled_with_features.csv')
df_l.head()

df = pd.read_csv('labeled_with_features.csv')
df = df.drop_duplicates(subset=['text', 'class_label'], keep=False)  # drop duplicates in dataset
df = df.drop(6322)
df['date'] =  pd.to_datetime(df['created_at'])
df['tweet_id'] = df['tweet_id'].astype(np.int64)
#print(df[df['created_at'].str.len() > 30])
df.head()

df1 = pd.read_csv('hurricane/hurricane_matthew_2016_train.tsv', delimiter='\t')
df1_1 = pd.read_csv('hurricane/hurricane_matthew_2016_dev.tsv', delimiter='\t')
df1_2 = pd.read_csv('hurricane/hurricane_matthew_2016_test.tsv', delimiter='\t')
df1 = pd.concat([df1, df1_1, df1_2])
df1['hurricane'] = 'matthew'
df1['fatalities'] = 603
df1['damage(billion USD)'] = 16.47
df1.head()

df2 = pd.read_csv('hurricane/hurricane_irma_2017_train.tsv', delimiter='\t')
df2_1 = pd.read_csv('hurricane/hurricane_irma_2017_dev.tsv', delimiter='\t')
df2_2 = pd.read_csv('hurricane/hurricane_irma_2017_test.tsv', delimiter='\t')
df2 = pd.concat([df2, df2_1, df2_2])
df2['hurricane'] = 'irma'
df2['fatalities'] = 52
df2['damage(billion USD)'] = 77.16
df2.head()

df3 = pd.read_csv('hurricane/hurricane_maria_2017_train.tsv', delimiter='\t')
df3_1 = pd.read_csv('hurricane/hurricane_maria_2017_test.tsv', delimiter='\t')
df3_2 = pd.read_csv('hurricane/hurricane_maria_2017_dev.tsv', delimiter='\t')
df3 = pd.concat([df3, df3_1, df3_2])
df3['hurricane'] = 'maria'
df3['fatalities'] = 2982
df3['damage(billion USD)'] = 90
df3.head()

df4 = pd.read_csv('hurricane/hurricane_harvey_2017_train.tsv', delimiter='\t')
df4_1 = pd.read_csv('hurricane/hurricane_harvey_2017_test.tsv', delimiter='\t')
df4_2 = pd.read_csv('hurricane/hurricane_harvey_2017_dev.tsv', delimiter='\t')
df4 = pd.concat([df4, df4_1, df4_2])
df4['hurricane'] = 'harvey'
df4['fatalities'] = 107
df4['damage(billion USD)'] = 125
df4.head()

df_h = pd.concat([df1, df2, df3, df4], ignore_index=True)
df_h = df_h.rename(columns={'tweet_text' : 'text'})
df_h

df_train = df_h.merge(df, on = ['tweet_id', 'class_label'],  how = 'inner')
#df_train.iloc[7235]
df_train = df_train.drop(7235)

df_mat = df_train[df_train['hurricane'] == 'matthew']
scaler = MinMaxScaler(feature_range=(0, 100))
df_mat['norm_date'] = scaler.fit_transform(df_mat['date'].values.reshape(-1,1))

df_irma = df_train[df_train['hurricane'] == 'irma']
df_irma['norm_date'] = scaler.fit_transform(df_irma['date'].values.reshape(-1,1))

df_maria = df_train[df_train['hurricane'] == 'maria']
df_maria['norm_date'] = scaler.fit_transform(df_maria['date'].values.reshape(-1,1))

df_harvey = df_train[df_train['hurricane'] == 'harvey']
df_harvey['norm_date'] = scaler.fit_transform(df_harvey['date'].values.reshape(-1,1))

df_new_train = pd.concat([df_mat, df_irma, df_maria, df_harvey], ignore_index=True)
len(df_new_train)

df_new_train = df_new_train.rename(columns={'text_x' : 'text'})
df_new_train = df_new_train.drop(columns=['text_y'])

df_new_train.to_csv('train_hurricane_with_name.csv')

scaler = MinMaxScaler(feature_range=(0, 100))

df_arthur = pd.read_csv('hurricane/CLEAN_ADD_HUR_DATA/arthur_2014_yn.csv')
df_arthur['hurricane'] = 'arthur'
df_arthur['fatalities'] = 2
df_arthur['damage(billion USD)'] = 0.0395
df_arthur['date'] =  pd.to_datetime(df_arthur['created_at'])
#df_arthur.iloc[208]
df_arthur = df_arthur.drop(208)
df_arthur['norm_date'] = scaler.fit_transform(df_arthur['date'].values.reshape(-1,1))

df_barry = pd.read_csv('hurricane/CLEAN_ADD_HUR_DATA/barry_2019_yn.csv')
df_barry['hurricane'] = 'barry'
df_barry['fatalities'] = 1
df_barry['damage(billion USD)'] = 0.6
df_barry['date'] =  pd.to_datetime(df_barry['created_at'])
df_barry['norm_date'] = scaler.fit_transform(df_barry['date'].values.reshape(-1,1))

df_florence = pd.read_csv('hurricane/CLEAN_ADD_HUR_DATA/florence_2018_yn.csv')
df_florence['hurricane'] = 'florence'
df_florence['fatalities'] = 53
df_florence['damage(billion USD)'] = 24.23
df_florence['date'] =  pd.to_datetime(df_florence['created_at'])
df_florence['norm_date'] = scaler.fit_transform(df_florence['date'].values.reshape(-1,1))

df_ian = pd.read_csv('hurricane/CLEAN_ADD_HUR_DATA/ian_2022_yn.csv')
df_ian['hurricane'] = 'ian'
df_ian['fatalities'] = 157
df_ian['damage(billion USD)'] = 50.2
df_ian['date'] =  pd.to_datetime(df_ian['created_at'])
df_ian['norm_date'] = scaler.fit_transform(df_ian['date'].values.reshape(-1,1))

df_isaias = pd.read_csv('hurricane/CLEAN_ADD_HUR_DATA/isaias_2020_yn.csv')
df_isaias['hurricane'] = 'isaias'
df_isaias['fatalities'] = 18
df_isaias['damage(billion USD)'] = 4.73
df_isaias['date'] =  pd.to_datetime(df_isaias['created_at'])
df_isaias['norm_date'] = scaler.fit_transform(df_isaias['date'].values.reshape(-1,1))

df_nicholas = pd.read_csv('hurricane/CLEAN_ADD_HUR_DATA/nicholas_2021_yn.csv')
df_nicholas['hurricane'] = 'nicholas'
df_nicholas['fatalities'] = 4
df_nicholas['damage(billion USD)'] = 1.1
df_nicholas['date'] =  pd.to_datetime(df_nicholas['created_at'])
df_nicholas['norm_date'] = scaler.fit_transform(df_nicholas['date'].values.reshape(-1,1))

df_sandy = pd.read_csv('hurricane/CLEAN_ADD_HUR_DATA/sandy_2012_yn.csv')
df_sandy['hurricane'] = 'sandy'
df_sandy['fatalities'] = 233
df_sandy['damage(billion USD)'] = 68.7
df_sandy['date'] =  pd.to_datetime(df_sandy['created_at'])
df_sandy['norm_date'] = scaler.fit_transform(df_sandy['date'].values.reshape(-1,1))

df_t = pd.concat([df_arthur, df_barry, df_florence, df_ian, df_isaias, df_nicholas, df_sandy], ignore_index=True)
df_t = df_t.rename(columns={'tweet_text' : 'text'})
df_t

df_t.to_csv('test_hurricane_with_name.csv')
